Add basic x86 instructions

+++ cpexp/source_related/c4e/memory.py
@@ -3,6 +3,6 @@ from cpexp.generic.memory import DataType
 
 class C4eType(DataType):
     types = {
-        'int': (1, 64),
+        'long': (1, 64),
         'float': (2, 64)
     }
